+++ mmadu-identity/src/main/java/com/mmadu/identity/services/token/TokenService.java
@@ -1,4 +1,4 @@
-package com.mmadu.identity.services;
+package com.mmadu.identity.services.token;
 
 import com.mmadu.identity.models.token.TokenRequest;
 import com.mmadu.identity.models.token.TokenResponse;
